Fix custom postfix attributes, and improve specs to cover them
The custom postfix TLS attributes - in particular the CA file path
- were not being applied because of the changes in postfix that
moved calculation of dynamic auth/tls attributes to recipes. This
is because postfix redefines those attributes as default, which
overrides our default attributes.

For now, our attributes have to be normal. It may however be
possible to remove some and refactor along the lines of postfix
cookbook.