Refactor postfix attribute config to take advantage postfix::_attributes
Postfix now compiles the various dependent configs - eg for sasl_auth
and TLS - in the attributes recipe rather than in its own attrribute
definitions. This allows us to simplify our own attributes as postfix
will build the relevant dynamic config for us.

Left the specs in to avoid future unpleasantness if the postfix
defaults change, so this won't affect default behaviour but makes
it simpler to override.