Fix breaking change to postfix sasl_auth configuration in postfix > 5.3.0
Per chef-cookbooks/postfix#134, the 5.3.0 release has a non-advertised
breaking change that requires different config attributes for the sasl_auth
for remote relays. @PeterGrace fixed this in his fork : merged back here with
an additional sanity check to throw if (unexpectedly) a project defines an
explicit value for these config attributes.